requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
asyncpg
sqlalchemy[asyncio]
pandas
openpyxl
python-dotenv
loguru
aiohttp
